requirements.txt
YooMoney==0.1.0
aiogram==2.25.1
asyncpg==0.27.0
aiohttp==3.8.4
certifi==2022.12.7
ujson==5.7.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pytz==2023.3
requests==2.30.0
urllib3==2.0.2
yarl==1.9.2
python-dateutil==2.2